token_safety/requirements.txt
requests>=2.31.0
web3>=6.15.0
supabase>=2.3.0
python-dotenv>=1.0.0
